fix cash

+++ sdk/src/main/java/ru/evotor/sdk/payment/entities/CashPaymentBody.kt
@@ -11,7 +11,7 @@ data class CashPaymentBody(
     /**
      * Сумма платежа
      */
-    var amount: BigDecimal? = null,
+    var amount: Long? = null,
     /**
      * Описание платежа
      */
